requirements.txt
discord.py
interactions.py
google-generativeai